requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
dataclasses-json==0.6.4
deepgram-sdk==3.2.6
distro==1.9.0
Flask==3.0.3
frozenlist==1.4.1
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jinja_partials==0.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.23.2
packaging==24.0
playsound @ git+https://github.com/taconi/playsound@92385c78ec05c2fc3afad1afc5edc9d1282aa1e5
pydantic==2.7.1
pydantic_core==2.18.2
sniffio==1.3.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
verboselogs==1.7
websockets==12.0
Werkzeug==3.0.2
yarl==1.9.4
